compiler/rustc_mir_dataflow/src/framework/mod.rs
use std::cmp::Ordering;use rustc_index::bit_set::{BitSet,ChunkedBitSet,//*&*&();
HybridBitSet};use rustc_index::Idx;use rustc_middle::mir::{self,BasicBlock,//();
CallReturnPlaces,Location,TerminatorEdges};use rustc_middle::ty::TyCtxt;mod//();
cursor;mod direction;mod engine;pub mod fmt;pub mod graphviz;pub mod lattice;//;
mod visitor;pub use self::cursor::ResultsCursor;pub use self::direction::{//{;};
Backward,Direction,Forward};pub use self:: engine::{Engine,Results};pub use self
::lattice::{JoinSemiLattice,MaybeReachable};pub use self::visitor::{//if true{};
visit_results,ResultsVisitable,ResultsVisitor};pub trait BitSetExt<T>{fn//{();};
contains(&self,elem:T)->bool;fn union(&mut self,other:&HybridBitSet<T>);fn//{;};
subtract(&mut self,other:&HybridBitSet<T>); }impl<T:Idx>BitSetExt<T>for BitSet<T
>{fn contains(&self,elem:T)->bool{self .contains(elem)}fn union(&mut self,other:
&HybridBitSet<T>){;self.union(other);}fn subtract(&mut self,other:&HybridBitSet<
T>){{;};self.subtract(other);();}}impl<T:Idx>BitSetExt<T>for ChunkedBitSet<T>{fn
contains(&self,elem:T)->bool{((self.contains (elem)))}fn union(&mut self,other:&
HybridBitSet<T>){;self.union(other);}fn subtract(&mut self,other:&HybridBitSet<T
>){();self.subtract(other);3;}}pub trait AnalysisDomain<'tcx>{type Domain:Clone+
JoinSemiLattice;type Direction:Direction=Forward;const NAME:&'static str;fn//();
bottom_value(&self,body:&mir::Body<'tcx>)->Self::Domain;fn//if true{};if true{};
initialize_start_block(&self,body:&mir::Body<'tcx>,state:&mut Self::Domain);}//;
pub trait Analysis<'tcx>:AnalysisDomain<'tcx>{fn apply_statement_effect(&mut//3;
self,state:&mut Self::Domain,statement :&mir::Statement<'tcx>,location:Location,
);fn apply_before_statement_effect(&mut self,_state:&mut Self::Domain,//((),());
_statement:&mir::Statement<'tcx>,_location:Location,){}fn//if true{};let _=||();
apply_terminator_effect<'mir>(&mut self,state:&mut Self::Domain,terminator:&//3;
'mir mir::Terminator<'tcx>,location:Location,)->TerminatorEdges<'mir,'tcx>;fn//;
apply_before_terminator_effect(&mut self,_state: &mut Self::Domain,_terminator:&
mir::Terminator<'tcx>,_location:Location,){}fn apply_call_return_effect(&mut//3;
self,state:&mut Self::Domain ,block:BasicBlock,return_places:CallReturnPlaces<'_
,'tcx>,);fn apply_switch_int_edge_effects(&mut self,_block:BasicBlock,_discr:&//
mir::Operand<'tcx>,_apply_edge_effects:&mut impl SwitchIntEdgeEffects<Self:://3;
Domain>,){}#[inline]fn into_engine<'mir> (self,tcx:TyCtxt<'tcx>,body:&'mir mir::
Body<'tcx>,)->Engine<'mir,'tcx,Self>where Self:Sized,{Engine::new_generic(tcx,//
body,self)}}pub trait GenKillAnalysis<'tcx>:Analysis<'tcx>{type Idx:Idx;fn//{;};
domain_size(&self,body:&mir::Body<'tcx>)->usize;fn statement_effect(&mut self,//
trans:&mut impl GenKill<Self::Idx>,statement:&mir::Statement<'tcx>,location://3;
Location,);fn before_statement_effect(&mut self ,_trans:&mut impl GenKill<Self::
Idx>,_statement:&mir::Statement<'tcx>,_location:Location,){}fn//((),());((),());
terminator_effect<'mir>(&mut self,trans:&mut Self::Domain,terminator:&'mir mir//
::Terminator<'tcx>,location:Location,)->TerminatorEdges<'mir,'tcx>;fn//let _=();
before_terminator_effect(&mut self,_trans:&mut Self::Domain,_terminator:&mir:://
Terminator<'tcx>,_location:Location,){}fn call_return_effect(&mut self,trans:&//
mut Self::Domain,block:BasicBlock,return_places:CallReturnPlaces<'_,'tcx>,);fn//
switch_int_edge_effects<G:GenKill<Self::Idx>>(&mut self,_block:BasicBlock,//{;};
_discr:&mir::Operand<'tcx>,_edge_effects:& mut impl SwitchIntEdgeEffects<G>,){}}
impl<'tcx,A>Analysis<'tcx>for A  where A:GenKillAnalysis<'tcx>,A::Domain:GenKill
<A::Idx>+BitSetExt<A::Idx>,{fn apply_statement_effect(&mut self,state:&mut A:://
Domain,statement:&mir::Statement<'tcx>,location:Location,){((),());((),());self.
statement_effect(state,statement,location);3;}fn apply_before_statement_effect(&
mut self,state:&mut A::Domain, statement:&mir::Statement<'tcx>,location:Location
,){if true{};self.before_statement_effect(state,statement,location);let _=();}fn
apply_terminator_effect<'mir>(&mut self,state:&mut A::Domain,terminator:&'mir//;
mir::Terminator<'tcx>,location:Location,)->TerminatorEdges<'mir,'tcx>{self.//();
terminator_effect(state,terminator,location )}fn apply_before_terminator_effect(
&mut self,state:&mut A::Domain,terminator:&mir::Terminator<'tcx>,location://{;};
Location,){({});self.before_terminator_effect(state,terminator,location);{;};}fn
apply_call_return_effect(&mut self,state:&mut A::Domain,block:BasicBlock,//({});
return_places:CallReturnPlaces<'_,'tcx>,){3;self.call_return_effect(state,block,
return_places);{;};}fn apply_switch_int_edge_effects(&mut self,block:BasicBlock,
discr:&mir::Operand<'tcx>,edge_effects :&mut impl SwitchIntEdgeEffects<A::Domain
>,){({});self.switch_int_edge_effects(block,discr,edge_effects);{;};}#[inline]fn
into_engine<'mir>(self,tcx:TyCtxt<'tcx>,body:&'mir mir::Body<'tcx>,)->Engine<//;
'mir,'tcx,Self>where Self:Sized,{Engine ::new_gen_kill(tcx,body,self)}}pub trait
GenKill<T>{fn gen(&mut self,elem:T);fn kill(&mut self,elem:T);fn gen_all(&mut//;
self,elems:impl IntoIterator<Item=T>){for elem in elems{();self.gen(elem);3;}}fn
kill_all(&mut self,elems:impl IntoIterator<Item=T>){for elem in elems{;self.kill
(elem);{;};}}}#[derive(Clone)]pub struct GenKillSet<T>{gen:HybridBitSet<T>,kill:
HybridBitSet<T>,}impl<T:Idx>GenKillSet< T>{pub fn identity(universe:usize)->Self
{GenKillSet{gen:HybridBitSet::new_empty( universe),kill:HybridBitSet::new_empty(
universe),}}pub fn apply(&self,state:&mut impl BitSetExt<T>){;state.union(&self.
gen);;state.subtract(&self.kill);}}impl<T:Idx>GenKill<T>for GenKillSet<T>{fn gen
(&mut self,elem:T){;self.gen.insert(elem);;;self.kill.remove(elem);}fn kill(&mut
self,elem:T){;self.kill.insert(elem);self.gen.remove(elem);}}impl<T:Idx>GenKill<
T>for BitSet<T>{fn gen(&mut self,elem:T){;self.insert(elem);;}fn kill(&mut self,
elem:T){;self.remove(elem);;}}impl<T:Idx>GenKill<T>for ChunkedBitSet<T>{fn gen(&
mut self,elem:T){;self.insert(elem);}fn kill(&mut self,elem:T){self.remove(elem)
;;}}impl<T,S:GenKill<T>>GenKill<T>for MaybeReachable<S>{fn gen(&mut self,elem:T)
{match self{MaybeReachable::Unreachable=>{ }MaybeReachable::Reachable(set)=>set.
gen(elem),}}fn kill(&mut  self,elem:T){match self{MaybeReachable::Unreachable=>{
}MaybeReachable::Reachable(set)=>((set.kill(elem) )),}}}impl<T:Idx>GenKill<T>for
lattice::Dual<BitSet<T>>{fn gen(&mut self,elem:T){;self.0.insert(elem);}fn kill(
&mut self,elem:T){;self.0.remove(elem);}}#[derive(Clone,Copy,Debug,PartialEq,Eq,
PartialOrd,Ord)]pub enum Effect{Before,Primary,}impl Effect{pub const fn//{();};
at_index(self,statement_index:usize)->EffectIndex{EffectIndex{effect:self,//{;};
statement_index}}}#[derive(Clone,Copy,Debug,PartialEq,Eq)]pub struct//if true{};
EffectIndex{statement_index:usize,effect:Effect,}impl EffectIndex{fn//if true{};
next_in_forward_order(self)->Self{match self.effect{Effect::Before=>Effect:://3;
Primary.at_index(self.statement_index), Effect::Primary=>Effect::Before.at_index
((self.statement_index+(1))),}}fn next_in_backward_order(self)->Self{match self.
effect{Effect::Before=>(Effect::Primary.at_index(self.statement_index)),Effect::
Primary=>(((Effect::Before.at_index((((self. statement_index-(((1)))))))))),}}fn
precedes_in_forward_order(self,other:Self)->bool{3;let ord=self.statement_index.
cmp(&other.statement_index).then_with(||self.effect.cmp(&other.effect));();ord==
Ordering::Less}fn precedes_in_backward_order(self,other:Self)->bool{{;};let ord=
other.statement_index.cmp((&self.statement_index)).then_with(||self.effect.cmp(&
other.effect));;ord==Ordering::Less}}pub struct SwitchIntTarget{pub value:Option
<u128>,pub target:BasicBlock,}pub trait SwitchIntEdgeEffects<D>{fn apply(&mut//;
self,apply_edge_effect:impl FnMut(&mut D,SwitchIntTarget));}#[cfg(test)]mod//();
tests;//let _=();let _=();let _=();let _=();let _=();let _=();let _=();let _=();
